ETL Manager Database Script
This script will;
- Create an ETLManager database on your target MS SQL server.
- Create an ETLManager table within the newly created ETLManager database on your chosen MS SQL server
- Create a LoadConfig stored procedure within the newly created ETLManager database on your chosen MS SQL server

To use this script, connect to your Microsoft SQL server of choice and execute the script.
This script has been tested from within Visual Studio 2017 but should run equally well in Query Analyzer.

This is an example project, thus this script is ridiculously simple. As a result it should be compatible with most older version MS SQL servers servers.